login integration and folder structure
