+++ package.json
@@ -18,10 +18,10 @@
     "@react-navigation/native-stack": "^6.9.26",
     "@reduxjs/toolkit": "^2.2.5",
     "axios": "^1.7.2",
+    "expo-linear-gradient": "^13.0.2",
     "formik": "^2.4.6",
     "i18next": "^23.11.5",
     "jwt-decode": "^4.0.0",
-    "nativewind": "^4.0.1",
     "react": "18.2.0",
     "react-i18next": "^14.1.2",
     "react-native": "0.74.2",
@@ -29,7 +29,6 @@
     "react-native-date-picker": "^5.0.4",
     "react-native-gesture-handler": "^2.16.2",
     "react-native-gifted-charts": "^1.4.11",
-    "react-native-linear-gradient": "^2.8.3",
     "react-native-modal": "^13.0.1",
     "react-native-orientation-locker": "^1.7.0",
     "react-native-reanimated": "^3.12.1",
