--- src/components/StudentInfo.js
@@ -1,10 +1,0 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-export default function StudentInfo() {
-  return (
-    <View>
-      <Text>StudentInfo</Text>
-    </View>
-  )
-}
