+++ src/screens/Login/Container/index.js
@@ -34,7 +34,7 @@ export default function LoginScreen() {
           <TouchableOpacity
             style={styles.swipeContainer}
             onPress={handleOpenPress}>
-            <Text style={styles.swipeText}>Click to Login</Text>
+            <Text style={styles.swipeText}>Tap to Login</Text>
           </TouchableOpacity>
           <BottomSheet
             enablePanDownToClose={true}
