+++ src/screens/attendance/components/card/styles.js
@@ -1,0 +1,101 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  cancelButton: {
+    position: 'absolute',
+    top: 20,
+    right: 20,
+    padding: 10,
+    borderRadius: 30,
+    backgroundColor: '#4E2973',
+  },
+  cancelButtonText: {
+    color: 'white',
+    fontSize: 16,
+    fontFamily: 'Satoshi',
+  },
+  card: {
+    alignSelf: 'center',
+    position: 'absolute',
+    top: 90,
+    borderRadius: 30,
+    overflow: 'hidden',
+  },
+  image: {
+    width: '100%',
+    height: '100%',
+    borderRadius: 30,
+  },
+  overlay: {
+    ...StyleSheet.absoluteFillObject,
+  },
+  startAttendanceContainer: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    borderRadius: 30,
+    backgroundColor: 'rgba(255,255,255,0.6)',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  startAttendanceButton: {
+    position: 'absolute',
+    // top: '50%',
+    // left: '50%',
+    // transform: [{translateX: -50}, {translateY: -50}],
+    borderRadius: 40,
+    backgroundColor: '#4E2973',
+    paddingHorizontal: 5,
+    paddingVertical: 2,
+    height: 50,
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  startAttendanceButtonText: {
+    color: 'white',
+    fontSize: 18,
+    fontFamily: 'Satoshi',
+    paddingHorizontal: 10,
+  },
+  nameText: {
+    position: 'absolute',
+    bottom: 100,
+    left: 20,
+    color: '#fff',
+    fontSize: 28,
+    fontWeight: '700',
+    fontFamily: 'Satoshi',
+  },
+  detailText1: {
+    position: 'absolute',
+    bottom: 65,
+    left: 20,
+    color: '#fff',
+    fontSize: 18,
+    fontWeight: '400',
+    fontFamily: 'Satoshi',
+  },
+  detailText2: {
+    position: 'absolute',
+    bottom: 30,
+    left: 20,
+    color: '#fff',
+    fontSize: 18,
+    fontWeight: '400',
+    fontFamily: 'Satoshi',
+  },
+  attendanceChoiceContainer: {
+    position: 'absolute',
+    top: 60,
+  },
+  attendanceChoiceText: {
+    fontSize: 40,
+    borderWidth: 4,
+    paddingLeft: 10,
+    paddingRight: 10,
+    fontFamily: 'Satoshi',
+  },
+});
