+++ src/screens/attendance/container/index.js
@@ -1,9 +1,9 @@
 import {View, Text} from 'react-native';
 import React, {useContext, useEffect, useState} from 'react';
-import AttendancePage from '../components/AttendancePage';
-import AttendanceCompleted from '../components/AttendanceCompleted';
-import {axiosClient} from '../services/axiosClient';
-import {AuthContext} from '../context/AuthContext';
+import AttendancePage from '@src/screens/attendance/components/attendancePage/index.js';
+import AttendanceCompleted from '@src/screens/attendance/components/completed/index.js';
+import {axiosClient} from '@src/services/axiosClient';
+import {AuthContext} from '@src/context/AuthContext';
 
 export default function AttendanceScreen({setAttendanceStarted}) {
   const [isTodayAttendance, setIsTodayAttendance] = useState();
@@ -14,6 +14,7 @@ export default function AttendanceScreen({setAttendanceStarted}) {
       const res = await axiosClient.get(
         `attendance/check-attendance-marked/${SectionId}`,
       );
+      // console.log(res.data);
       if (res.data.result) {
         setIsTodayAttendance(false);
       } else {
